AudioFlinger: fix createEffect() for second client attached
The refactoring of AudioFlinger::createEffect() with AIDL (commit ce182940)
introduced a regression for the case where a second client
connects to an already existing effect.
In this case, the conventional return status of ALREADY_EXISTS is
interpreted as an error by the AIDL infrastructure and the response
data is dropped during the transaction causing the effect creation
to fail.

Test: atest audioeffect_tests:DefaultEffectTest#ManageSourceDefaultEffects

Change-Id: I858dd0ee2a200901ec229d4a5cabdf4a458ee88c
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>